Add rudimentary happy path tests
