+++ package.json
@@ -67,6 +67,7 @@
 		"eslint-plugin-regexp": "^2.1.1",
 		"eslint-plugin-vitest": "^0.3.9",
 		"eslint-plugin-yml": "^1.10.0",
+		"happy-dom": "^12.10.3",
 		"husky": "^8.0.3",
 		"jsonc-eslint-parser": "^2.4.0",
 		"knip": "^2.41.0",
