+++ src/content-script.ts
@@ -2,6 +2,7 @@ import * as yaml from "js-yaml";
 import Mustache from "mustache";
 
 import { createElement } from "./elements.js";
+import { getSoon } from "./getSoon.js";
 import { isBodyWithReplies, isRepositoryDetails } from "./validations.js";
 
 // TODO: Add handling for a rejection
@@ -73,23 +74,6 @@ async function main() {
 	if (!newCommentField) {
 		console.error("Couldn't find comment field");
 		return;
-	}
-
-	async function getSoon<Value extends Node | NodeList | null>(
-		getter: () => Value,
-	) {
-		for (let i = 0; i < 100; i += 1) {
-			const result = getter();
-			if (result && (!(result instanceof NodeList) || result.length)) {
-				return result;
-			}
-
-			await new Promise((resolve) => setTimeout(resolve, i));
-		}
-
-		// TODO: Add handling for a rejection
-		// https://github.com/JoshuaKGoldberg/refined-saved-replies/issues/2
-		throw new Error(`Element was never found 😔`);
 	}
 
 	const onOpenSavedRepliesButtonClick = async () => {
